pyproject.toml
[project]
name = "comptroller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=1.2.8",
    "langchain-anthropic>=1.3.1",
    "langchain-openai>=1.1.7",
    "tiktoken>=0.12.0",
]
